backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
pydantic==2.10.1
pydantic-settings==2.6.1
uvicorn[standard]==0.32.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
psycopg2-binary==2.9.10
python-dotenv==1.0.1
alembic==1.14.0
asyncpg==0.30.0
uvicorn